requirements.txt
aioredis==1.3.1
amqp==5.0.6
argon2-cffi==21.1.0
arrow==1.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.18.39
botocore==1.21.40
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
daphne==3.0.2
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.8
django-active-link==0.1.8
django-appconf==1.0.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.13.3
django-ratelimit==3.0.1
django-recaptcha==2.0.6
django-redis==5.0.0
django-redis-cache==3.0.0
django-storages==1.11.1
django-taggit==1.5.1
django-timezone-field==4.2.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.3
markdown2==2.4.1
msgpack==1.0.2
pilkit==2.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyquery==1.4.3
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
readtime==1.1.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.interface==5.4.0
